Markdown snippet files and multimethod split-and-to-html.
A first test for markdown snippets worked. It was necessary to change
md-files to look only in src/pages and src/posts for markdown files.
This prevents all-pages from reading snippets and from crashing due to
missing meta data.

There is a first sketch for a multimethod that splits files and converts
them to html. This is neither used nor tested!