Separate namespaces for rendering, code and api; some refactorings.
There is now a functioning namespace for the API (ecstatic.api), the
rendering of templates (ecstatic.render) and the loading of the custom code
(ecstatic.code). The latter ones refer to  ecstatic.api.

To fix cyclic dependencies but to be sure that the aforementioned namespaces
exist at the time the templates get rendered, the -main function now resides
in its own namespace ecstatic.main that depends on ecstatic.core and ecstatic.api.

The refactorings are:
 - change the name of the dynamic variable in to *in-dir* to be more descriptive
   and to match the naming conventions.
 - refactor the code to render hiccup data into its own function and use it throughout the
   program